% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{oscorad2easi}
\alias{oscorad2easi}
\title{Map OSCORAD to EASI}
\usage{
oscorad2easi(x, type = "lm")
}
\arguments{
\item{x}{a vector of OSCORAD values}

\item{type}{model type used for mapping, defaults to ‘"nlm"’}
}
\value{
a vector of predicted EASI values
}
\description{
Map OSCORAD to EASI using models of varying complexity
}
\details{
Depending on the choice of ‘type’, 
a linear model ('"lm"', lm(y~x)), 
a non-linear model ('"nlm"', nls(y~b0+b1⋅x+b2⋅x2)) or 
locally estimated scatterplot smoothing ('"loess"', loess(y~x)) 
is used to predict EASI.
}
\examples{
x<-sample(0:83, size=10, replace=T)
y<-oscorad2easi(x=x, type="nls")
}
\references{
Hübenthal M, Dai C et al..
“Mapping OSCORAD and SCORAD to EASI to facilitate large-scale meta-analyses of omics data”
(in preparation)
}
\author{
Matthias Hübenthal <<mhuebenthal@dermatology.uni-kiel.de>>, 
Chengliang Dai <<chengliang.dai@ucb.com>> 
et al.
}
